Tcl interp cannot be accessed by different threads (see http://www.tcl.tk/doc/howto/thread_model.html) while previous code did it and cause break.
So now events will be dispatched via a Tcl command
you've to invoke by yourself: iaxcGetEvents. This
is a kind of polling mechanism so it's up to you
decide how often to look for events (if you ever
need to do it).

To avoid sync problems between different threads,
each access to the event list has been regulated
via mutex lock/unlock.